separate LocationMap to allow for NoSSR
