+++ packages/experimental-renderers/package.json
@@ -63,6 +63,7 @@
     "classnames": "^2.3.2",
     "leaflet": "^1.9.4",
     "lodash": "^4.17.21",
+    "next": "^13.4.19",
     "react-leaflet": "^4.2.1"
   }
 }
